requirements.txt
# Core scientific stack
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
pyyaml
